+++ modules/linkformat_test.go
@@ -1,4 +1,4 @@
-package renter
+package modules
 
 import (
 	"testing"
@@ -26,8 +26,7 @@ func TestLinkFormat(t *testing.T) {
 	}
 	if ldDecoded != ld {
 		t.Error("encoded data and decoded data do not match")
+		t.Error(ld)
+		t.Error(ldDecoded)
 	}
-
-	// TODO: Loop to test a bunch of different random inputs to make sure things
-	// are compatible across all sorts of inputs.
 }