+++ modules/renter/linkfile_test.go
@@ -9,10 +9,15 @@ import (
 func TestLinkfileLayoutEncoding(t *testing.T) {
 	// Try encoding an decoding a simple example.
 	llOriginal := linkfileLayout{
-		filesize:           1e6,
-		metadataSize:       14e3,
-		fanoutDataPieces:   8,
-		fanoutParityPieces: 3,
+		version:                 LinkfileVersion,
+		filesize:                1e6,
+		metadataSize:            14e3,
+		intraSectorDataPieces:   8,
+		intraSectorParityPieces: 3,
+		fanoutHeaderSize:        75e3,
+		fanoutExtensionSize:     9e9,
+		fanoutDataPieces:        10,
+		fanoutParityPieces:      20,
 	}
 	encoded := llOriginal.encode()
 	var llRecovered linkfileLayout
@@ -20,6 +25,4 @@ func TestLinkfileLayoutEncoding(t *testing.T) {
 	if llOriginal != llRecovered {
 		t.Fatal("encoding and decoding of linkfileLayout does not match")
 	}
-
-	// TODO: Try a wider range of values with randomness / fuzzing.
 }