+++ node/api/renter.go
@@ -252,7 +252,7 @@ type (
 	// RenterLinkfileHandlerPOST is the response that the api returns after the
 	// /renter/linkfile/ POST endpoint has been used.
 	RenterLinkfileHandlerPOST struct {
-		Sialink string `json:"sialink"`
+		Sialink modules.Sialink `json:"sialink"`
 	}
 )
 
@@ -286,7 +286,7 @@ func trimSiaDirFolderOnFiles(fis ...modules.FileInfo) (_ []modules.FileInfo, err
 	for i := range fis {
 		fis[i].SiaPath, err = fis[i].SiaPath.Rebase(modules.UserSiaPath(), modules.RootSiaPath())
 		if err != nil {
-			return nil, err
+			return nil, errors.AddContext(err, "unable to trim the user sia path from a provided fileinfo")
 		}
 	}
 	return fis, nil
@@ -1341,27 +1341,53 @@ func (api *API) renterRenameHandler(w http.ResponseWriter, req *http.Request, ps
 
 // renterFileHandler handles GET requests to the /renter/file/:siapath API endpoint.
 func (api *API) renterFileHandlerGET(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
+	// Determine the siapath that hte user wants to get the file from.
 	siaPath, err := modules.NewSiaPath(ps.ByName("siapath"))
 	if err != nil {
 		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
 		return
 	}
-	siaPath, err = rebaseInputSiaPath(siaPath)
-	if err != nil {
-		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
-		return
+
+	// Determine whether the user is requesting a user siapath, or a root
+	// siapath.
+	var root bool
+	rootStr := req.FormValue("root")
+	if rootStr != "" {
+		root, err = strconv.ParseBool(rootStr)
+		if err != nil {
+			WriteError(w, Error{"unable to parse 'root' arg"}, http.StatusBadRequest)
+			return
+		}
 	}
+
+	// Rebase the users input to the user folder if the user is requesting a
+	// user siapath.
+	if !root {
+		siaPath, err = rebaseInputSiaPath(siaPath)
+		if err != nil {
+			WriteError(w, Error{err.Error()}, http.StatusBadRequest)
+			return
+		}
+	}
+
+	// Fetch the file.
 	file, err := api.renter.File(siaPath)
 	if err != nil {
 		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
 		return
 	}
-	files, err := trimSiaDirFolderOnFiles(file)
-	if err != nil {
-		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
-		return
+
+	// If the user requested the user siapath, trim the dir folder so that the
+	// output is all centered around the user's folder.
+	if !root {
+		files, err := trimSiaDirFolderOnFiles(file)
+		if err != nil {
+			WriteError(w, Error{err.Error()}, http.StatusBadRequest)
+			return
+		}
+		file = files[0]
 	}
-	file = files[0]
+
 	WriteJSON(w, RenterFile{
 		File: file,
 	})
@@ -1691,9 +1717,7 @@ func parseDownloadParameters(w http.ResponseWriter, req *http.Request, ps httpro
 // renterSialinkHandlerGET accepts a sialink as input and will stream the data
 // from the sialink out of the response body as output.
 func (api *API) renterSialinkHandlerGET(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
-	// TODO: Add support for offset + len.
-
-	sialink := ps.ByName("sialink")
+	sialink := modules.Sialink(ps.ByName("sialink"))
 	metadata, data, err := api.renter.DownloadSialink(sialink)
 	if err != nil {
 		WriteError(w, Error{fmt.Sprintf("failed to fetch sialink: %v", err)}, http.StatusInternalServerError)
@@ -1708,7 +1732,7 @@ func (api *API) renterSialinkHandlerGET(w http.ResponseWriter, req *http.Request
 func (api *API) renterLinkfileHandlerPOST(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
 	// Parse out the intended siapath.
 	siaPathStr := ps.ByName("siapath")
-	siaPath, err := modules.UserSiaPath().Join(siaPathStr)
+	siaPath, err := modules.LinkfileSiaFolder.Join(siaPathStr)
 	if err != nil {
 		WriteError(w, Error{"invalid siapath provided: " + err.Error()}, http.StatusBadRequest)
 		return
@@ -1722,32 +1746,58 @@ func (api *API) renterLinkfileHandlerPOST(w http.ResponseWriter, req *http.Reque
 	}
 
 	// Check whether existing file should be overwritten
-	overwriteExistingFile := false
-	if f := queryForm.Get("overwriteexistingfile"); f != "" {
-		overwriteExistingFile, err = strconv.ParseBool(f)
+	force := false
+	if f := queryForm.Get("force"); f != "" {
+		force, err = strconv.ParseBool(f)
 		if err != nil {
-			WriteError(w, Error{"unable to parse 'overwriteexistingfile' parameter: " + err.Error()}, http.StatusBadRequest)
+			WriteError(w, Error{"unable to parse 'force' parameter: " + err.Error()}, http.StatusBadRequest)
 			return
 		}
 	}
 
-	// TODO: Erasure coding params - both for the base file and for the fanout.
+	// Check whether the redundancy has been set.
+	redundancy := uint8(0)
+	if rStr := queryForm.Get("redundancy"); rStr != "" {
+		if _, err := fmt.Sscan(rStr, &redundancy); err != nil {
+			WriteError(w, Error{"unable to parse redundancy: " + err.Error()}, http.StatusBadRequest)
+			return
+		}
+	}
 
 	// Call the renter to upload the linkfile and create a sialink.
 	name := queryForm.Get("name")
 	modeStr := queryForm.Get("mode")
-	var mode uint32
-	if modeStr == "" {
-		mode = modules.DefaultFilePerm
-	} else {
-		// TODO: parse modeStr instead of using default
-		mode = modules.DefaultFilePerm
+	var mode os.FileMode
+	if modeStr != "" {
+		_, err := fmt.Sscanf(modeStr, "%o", &mode)
+		if err != nil {
+			WriteError(w, Error{fmt.Sprintf("failed to parse file mode: %v", err)}, http.StatusBadRequest)
+			return
+		}
+	}
+	createTimeStr := queryForm.Get("createtime")
+	var createTime int64
+	if createTimeStr != "" {
+		_, err := fmt.Sscan(createTimeStr, &createTime)
+		if err != nil {
+			WriteError(w, Error{fmt.Sprintf("failed to parse file create time: %v", err)}, http.StatusBadRequest)
+			return
+		}
 	}
 	lfm := modules.LinkfileMetadata{
-		Name: name,
-		Mode: mode,
+		Name:       name,
+		Mode:       mode,
+		CreateTime: createTime,
 	}
-	sialink, err := api.renter.UploadLinkfile(lfm, siaPath, overwriteExistingFile, req.Body)
+	lup := modules.LinkfileUploadParameters{
+		SiaPath:             siaPath,
+		Force:               force,
+		BaseChunkRedundancy: redundancy,
+
+		FileMetadata: lfm,
+		Reader:       req.Body,
+	}
+	sialink, err := api.renter.UploadLinkfile(lup)
 	if err != nil {
 		WriteError(w, Error{fmt.Sprintf("failed to upload linkfile: %v", err)}, http.StatusBadRequest)
 		return
@@ -2050,6 +2100,7 @@ func (api *API) renterUploadStreamHandler(w http.ResponseWriter, req *http.Reque
 		ErasureCode: ec,
 		Force:       force,
 		Repair:      repair,
+		CipherType:  crypto.TypeDefaultRenter,
 	}
 	err = api.renter.UploadStreamFromReader(up, req.Body)
 	if err != nil {
@@ -2074,6 +2125,18 @@ func (api *API) renterValidateSiaPathHandler(w http.ResponseWriter, _ *http.Requ
 func (api *API) renterDirHandlerGET(w http.ResponseWriter, req *http.Request, ps httprouter.Params) {
 	var siaPath modules.SiaPath
 	var err error
+
+	// Check whether the user is requesting the directory from the root path.
+	var root bool
+	rootStr := req.FormValue("root")
+	if rootStr != "" {
+		root, err = strconv.ParseBool(rootStr)
+		if err != nil {
+			WriteError(w, Error{"unable to parse 'root' arg"}, http.StatusBadRequest)
+			return
+		}
+	}
+
 	str := ps.ByName("siapath")
 	if str == "" || str == "/" {
 		siaPath = modules.RootSiaPath()
@@ -2084,31 +2147,43 @@ func (api *API) renterDirHandlerGET(w http.ResponseWriter, req *http.Request, ps
 		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
 		return
 	}
-	siaPath, err = rebaseInputSiaPath(siaPath)
-	if err != nil {
-		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
-		return
+
+	if !root {
+		siaPath, err = rebaseInputSiaPath(siaPath)
+		if err != nil {
+			WriteError(w, Error{err.Error()}, http.StatusBadRequest)
+			return
+		}
 	}
+
 	directories, err := api.renter.DirList(siaPath)
 	if err != nil {
 		WriteError(w, Error{"failed to get directory contents:" + err.Error()}, http.StatusInternalServerError)
 		return
 	}
-	directories, err = trimSiaDirFolder(directories...)
-	if err != nil {
-		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
-		return
+
+	if !root {
+		directories, err = trimSiaDirFolder(directories...)
+		if err != nil {
+			WriteError(w, Error{err.Error()}, http.StatusBadRequest)
+			return
+		}
 	}
+
 	files, err := api.renter.FileList(siaPath, false, true)
 	if err != nil {
 		WriteError(w, Error{"failed to get file infos:" + err.Error()}, http.StatusInternalServerError)
 		return
 	}
-	files, err = trimSiaDirFolderOnFiles(files...)
-	if err != nil {
-		WriteError(w, Error{err.Error()}, http.StatusBadRequest)
-		return
+
+	if !root {
+		files, err = trimSiaDirFolderOnFiles(files...)
+		if err != nil {
+			WriteError(w, Error{err.Error()}, http.StatusBadRequest)
+			return
+		}
 	}
+
 	WriteJSON(w, RenterDirectory{
 		Directories: directories,
 		Files:       files,