+++ siatest/renter/linkfile_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"testing"
 
+	"gitlab.com/NebulousLabs/Sia/modules"
 	"gitlab.com/NebulousLabs/Sia/siatest"
 	"gitlab.com/NebulousLabs/fastrand"
 )
@@ -40,7 +41,30 @@ func TestLinkfile(t *testing.T) {
 	// Need it to be a reader.
 	reader := bytes.NewReader(data)
 	// Call the upload linkfile client call.
-	sialink, err := r.RenterLinkfilePost(reader, "testOne", "/home/user/linkfiles/testOne")
+	filename := "testOne"
+	uploadSiaPath, err := modules.NewSiaPath("testOnePath")
+	if err != nil {
+		t.Fatal(err)
+	}
+	// Quick fuzz on the force value so that sometimes it is set, sometimes it
+	// is not.
+	var force bool
+	if fastrand.Intn(2) == 0 {
+		force = true
+	}
+	lup := modules.LinkfileUploadParameters{
+		SiaPath:             uploadSiaPath,
+		Force:               force,
+		BaseChunkRedundancy: 2,
+		FileMetadata: modules.LinkfileMetadata{
+			Name:       filename,
+			Mode:       0600, // intentionally not the default
+			CreateTime: 1e6,  // intentionally before current time
+		},
+
+		Reader: reader,
+	}
+	sialink, err := r.RenterLinkfilePost(lup)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -56,4 +80,32 @@ func TestLinkfile(t *testing.T) {
 		t.Log(data)
 		t.Log(fetchedData)
 	}
+
+	// Check the metadata of the siafile, see that the metadata of the siafile
+	// has the sialink referenced.
+	linkfilePath, err := modules.LinkfileSiaFolder.Join(uploadSiaPath.String())
+	if err != nil {
+		t.Fatal(err)
+	}
+	renterFile, err := r.RenterFileRootGet(linkfilePath)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(renterFile.File.Sialinks) != 1 {
+		t.Fatal("expecting one sialink:", len(renterFile.File.Sialinks))
+	}
+	if renterFile.File.Sialinks[0] != sialink {
+		t.Error("sialinks should match")
+		t.Log(renterFile.File.Sialinks[0])
+		t.Log(sialink)
+	}
+
+	// TODO: Need to verify the mode, name, and create-time. At this time, I'm
+	// not sure how we can feed those out of the API. They aren't going to be
+	// the same as the siafile values, because the siafile was created
+	// separately.
+	//
+	// Maybe this can be accomplished by tagging a flag to the API which has the
+	// layout and metadata streamed as the first bytes? Maybe there is some
+	// easier way.
 }